+++ src/type-data-frame.c
@@ -437,6 +437,7 @@ SEXP df_cast(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg)
   // Restore data frame size before calling `vec_restore()`. `x` and
   // `to` might not have any columns to compute the original size.
   init_data_frame(out, size);
+  Rf_setAttrib(out, R_RowNamesSymbol, df_rownames(x));
 
   R_len_t extra_len = Rf_length(x) - common_len;
   if (extra_len) {